+++ software/sub_driver/src/time/Time.h
@@ -52,7 +52,7 @@
 
 template <
     std::size_t max_tasks = TIMER_MAX_TASKS,    /* max allocated tasks */
-    unsigned long long (*time_func)() = micros, /* time function for timer */
+    uint64_t (*time_func)() = micros, /* time function for timer */
     typename T = void *                         /* handler argument type */
     >
 class Timer
